test: consolidate closure and trait object e2e tests
