--- src/__tests__/fixtures/pipe-closure.ts
@@ -1,8 +1,0 @@
-export const pipeClosureVoyd = `
-use std::all
-
-pub fn main() -> i32
-  2
-    .(x: i32) => x + 3
-    .(x: i32) => x * 5
-`;
