--- src/__tests__/generic-trait.e2e.test.ts
@@ -1,20 +1,0 @@
-import { genericTraitVoyd } from "./fixtures/generic-trait.js";
-import { compile } from "../compiler.js";
-import { describe, test, beforeAll } from "vitest";
-import assert from "node:assert";
-import { getWasmFn, getWasmInstance } from "../lib/wasm.js";
-
-describe("E2E generic trait objects", () => {
-  let instance: WebAssembly.Instance;
-
-  beforeAll(async () => {
-    const mod = await compile(genericTraitVoyd);
-    instance = getWasmInstance(mod);
-  });
-
-  test("run returns correct value", (t) => {
-    const fn = getWasmFn("run", instance);
-    assert(fn, "Function exists");
-    t.expect(fn(), "run returns correct value").toEqual(1);
-  });
-});
